bridge-tsp-competition/pyproject.toml
[project]
name = "bridge-tsp-competition"
version = "1.0.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "MetaTrader5",
    "supabase==2.0.3",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
]
